Add lightweight volunteer onboarding workshop (#286)
* Add lightweight volunteer onboarding workshop

* Add learning objectives

* Reword exercise a little

* Focus on time limit a bit more
* Focus on desired outcomes of exchange

* Add to index

* Update common

* Revert "Add to index"

This reverts commit b32084a41f7eb4c6bd64d42bcb9b3b6f6fafce57.

The index relies on things already being merged.

* Create a placeholder 404 file with the correct permissions

Hugo will correctly replace this with the actual rendered 404 page. But
if this file doesn't already exist with the correct permissions, hugo
will for some reason which we can't debug create a placeholder file
with permissions 0444 instead of 0644, and then won't be able to
overwrite it with the correctly rendered file.